+++ R/analyze.R
@@ -5,6 +5,7 @@
 #'  \item{\link[=analyze.stanreg]{analyze.stanreg}}
 #'  \item{\link[=analyze.lmerModLmerTest]{analyze.merModLmerTest}}
 #'  \item{\link[=analyze.glmerMod]{analyze.glmerMod}}
+#'  \item{\link[=analyze.lm]{analyze.lm}}
 #'  }
 #'
 #' @param x object to analyze.